put files in folders
